+++ src/uniform.jl
@@ -109,6 +109,21 @@ _unique(A::AbstractUniformArray, ::Colon) = [value(A)]
 _unique(A::AbstractUniformArray, dim::Integer) =
     UniformArray(value(A), reduce_shape1(shape(A), dim))
 
+# Optimize mapping and broadcasting of functions for uniform arrays.
+Broadcast.broadcasted(f, A::AbstractUniformArray) = map(f, A)
+function Base.map(f, A::UniformArray)
+    val = f(value(A))
+    return UniformArray{typeof(val)}(BareBuild(), val, shape(A))
+end
+function Base.map(f, A::FastUniformArray)
+    val = f(value(A))
+    return FastUniformArray{typeof(val),ndims(A),val}(BareBuild(), shape(A))
+end
+function Base.map(f, A::MutableUniformArray)
+    val = f(value(A))
+    return MutableUniformArray{typeof(val)}(BareBuild(), val, shape(A))
+end
+
 # Optimize base reduction methods for uniform arrays.
 for func in (:all, :any,
              :minimum, :maximum, :extrema,